requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.7.0
openai==1.29.0
python-telegram-bot==21.1.1
requests~=2.31.0
tenacity==8.3.0
firebase-admin=6.5.0
google-cloud-firestore=2.16.1
anthropic=0.31.2
langchain-anthropic=0.1.20
langchain-core=0.2.22
